refactor: move mocked products data to separeted folder
