+++ products/getProductById.ts
@@ -1,6 +1,6 @@
 import type { APIGatewayProxyEvent } from 'aws-lambda';
 
-import { products } from './utils/products';
+import { products } from './mock/products';
 import prepareResponse from './utils/prepareResponse';
 
 export const handler = async (event: APIGatewayProxyEvent) => {
