--- test/nodejs-aws-shop-backend.test.ts
@@ -1,17 +1,0 @@
-// import * as cdk from 'aws-cdk-lib';
-// import { Template } from 'aws-cdk-lib/assertions';
-// import * as NodejsAwsShopBackend from '../lib/nodejs-aws-shop-backend-stack';
-
-// example test. To run these tests, uncomment this file along with the
-// example resource in lib/nodejs-aws-shop-backend-stack.ts
-test('SQS Queue Created', () => {
-//   const app = new cdk.App();
-//     // WHEN
-//   const stack = new NodejsAwsShopBackend.NodejsAwsShopBackendStack(app, 'MyTestStack');
-//     // THEN
-//   const template = Template.fromStack(stack);
-
-//   template.hasResourceProperties('AWS::SQS::Queue', {
-//     VisibilityTimeout: 300
-//   });
-});
